Unsuccessfully try html2canvas and jpdf npm modules to print bingo card as .pdf
This method did not work because html2canvas converts the html you want to print into a canvas element.
Since memes hosted on imgur.com are images that I want to export, it doesn't work because of cross-origin issues.
Going to try storing images in local storage and then print those to .pdf.